backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.8.2
pydantic-settings==2.4.0
python-dotenv==1.0.1
motor==3.5.1
pymongo[srv]==4.8.0
email-validator==2.2.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
pyjwt[crypto]==2.8.0
httpx==0.27.0
openai==1.35.13
websockets==12.0
